Инициализация таблицы voteAnswer

+++ src/main/java/com/javamentor/qa/platform/service/impl/TestDataInitService.java
@@ -130,7 +130,7 @@ public class TestDataInitService {
             answer.setIsHelpful(random.nextBoolean());
             answer.setIsDeletedByModerator(random.nextBoolean());
             answer.setQuestion(questionService.getById((long) (1 + (int) (Math.random() * 49))).get());
-            answer.setUser(userService.getById((long) (1 + (Math.random() * 49))).get());
+            answer.setUser(userService.getById((long) (1 + (int) (Math.random() * 49))).get());
             //добавдение модератора к рамдомной части ответов
             for (int y = 1 + (int) (Math.random() * 49); y <= x; y = y + 2) {
                 if (y == x) {
